Revert "deduplicate size_t functions"
This reverts commit 24941f530d1c1d115a28fb8f0a715a3d15b5f74f.

+++ File.h
@@ -111,10 +111,11 @@ public:
 
   // Write function
   void Write(const int v, const char *sep = NULL);
-  void Write(const size_t v, const char *sep = NULL);
+  void Write(const int v, const char *sep=NULL);
   void Write(const double v, const char *sep=NULL);
   void Write(const char *s);
-  void Write(const std::string& str);  
+  void Write(const std::string& str);
+  
 
 private:
